pyproject.toml
[tool.poetry]
name = "aoi"
license = "MIT"
version = "0.1.0"
description = "aoi"
readme = "README.md"
homepage = "https://github.com/ap-1/aoi"
repository = "https://github.com/ap-1/aoi.git"
authors = ["ap-1 <67872951+ap-1@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.15.0"
aiohttp = "^3.7.4"
asyncpraw = "^7.2.0"
"discord.py" = {extras = ["voice"], version = "^1.7.2"}
discord-py-slash-command = "^1.0.9"
youtube_dl = "^2021.4.26"
pyquery = "^1.4.3"

[tool.poetry.scripts]
start = "aoi.main:init"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/ap-1/aoi/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
